update test data with address key change
