+++ crates/core/src/client/helpers.rs
@@ -136,6 +136,8 @@ pub fn split_u256_into_field_elements(value: U256) -> [FieldElement; 2] {
 #[cfg(test)]
 mod tests {
 
+    use rstest::*;
+
     use super::*;
 
     #[test]
@@ -173,5 +175,34 @@ mod tests {
         let expected: Bytes =
             serde_json::from_str(include_str!("../models/test_data/bytecode/eth/counter.json")).unwrap();
         assert_eq!(expected, bytes);
+    }
+
+    #[rstest]
+    #[test]
+    #[case(
+        "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaabbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb",
+        "0xaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaabbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb"
+    )]
+    #[case(
+        "0x0000000000000000000000000000000000000000000000000000000000000000",
+        "0x0000000000000000000000000000000000000000000000000000000000000000"
+    )]
+    #[case(
+        "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff",
+        "0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"
+    )]
+    fn test_split_u256_into_field_elements(#[case] input: U256, #[case] expected: U256) {
+        // When
+        let result = split_u256_into_field_elements(input);
+
+        // Then
+        // Recalculate the U256 values using the resulting FieldElements
+        // The first is the low 128 bits of the U256 value
+        // The second is the high 128 bits of the U256 value and is left shifted by 128 bits
+        let result: U256 =
+            U256::from_be_bytes(result[1].to_bytes_be()) << 128 | U256::from_be_bytes(result[0].to_bytes_be());
+
+        // Assert that the expected and recombined U256 values are equal
+        assert_eq!(expected, result);
     }
 }
