+++ crates/core/src/models/test_data/conversion/starknet/event_invalid_key.json
@@ -1,13 +1,13 @@
 {
   "from_address": "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7",
   "keys": [
+    "0x2b61c43a85bd35987c5311215e8288b823a6873edeadbeef",
     "0x7f38d2db6e0da72afe22cefa3065bc63",
     "0x5998d146b8109b9444e9bb13ae9a548e",
     "0xa",
     "0x0",
     "0xb",
-    "0x0",
-    "0x2b61c43a85bd35987c5311215e8288b823a6873edeadbeef"
+    "0x0"
   ],
   "data": []
 }
