+++ crates/core/src/models/test_data/conversion/starknet/event_log4.json
@@ -1,6 +1,7 @@
 {
   "from_address": "0x049d36570d4e46f48e99674bd3fcc84644ddd6b96f7c741b1562b82f9e004dc7",
   "keys": [
+    "0x2b61c43a85bd35987c5311215e8288b823a6873e",
     "0xbf0a60652fafd614e9b6be2ca74a54a9",
     "0x8106949def8a44172f54941ce774c774",
     "0xa",
@@ -8,8 +9,7 @@
     "0xb",
     "0x0",
     "0xa",
-    "0x0",
-    "0x2b61c43a85bd35987c5311215e8288b823a6873e"
+    "0x0"
   ],
   "data": []
 }
